Cast get_knowledge oldest to int
Slack's conversations_history silently returns 0 messages when
`oldest` is passed as a float with sub-second precision
(e.g. "1774115668.4110897"). Unfiltered reads (days=None → "0")
worked; every days-filtered read returned empty. That made
get_status_overrides come back as NONE even with dozens of
relevant CORRECTION/FEEDBACK entries in the channel — the root
cause behind today's ADONIS/Mexico leaks.

Verified via local dry-run (test_pipeline.py): ADONIS no longer
appears, Mexico is correctly flagged as delayed in the team
summary, Josh's #1 is S3→UMES per the correction.

+++ knowledge.py
@@ -93,7 +93,11 @@ def get_knowledge(slack_client, entry_types=None, days=None):
     if not channel:
         return []
 
-    oldest = str(time.time() - (days * 86400)) if days else "0"
+    # Slack's conversations_history silently returns 0 messages when `oldest`
+    # is a float with sub-second precision like "1774115668.4110897". Cast to
+    # int — otherwise every days-filtered read returns empty, which is what
+    # made status overrides come back as NONE despite real corrections.
+    oldest = str(int(time.time() - (days * 86400))) if days else "0"
     entries = []
     cursor = None
